move organization query tests
